RDK-38587 - GTests for Xcast plugin
Reason for change:
Adding Gtests for Xcast plugin
TestProcedure: None
Risks: Low

+++ XCast/XCast.cpp
@@ -436,7 +436,7 @@ bool XCast::deleteFromDynamicAppCache(JsonArray applications)
     vector<string> appsToDelete;
     for (int iIndex = 0; iIndex < applications.Length(); iIndex++) {
         itrName = applications[iIndex].String();
-        LOGINFO("App name to delete: %s, size:%d", itrName.c_str(), strlen (itrName.c_str()));
+        LOGINFO("App name to delete: %s, size:%d", itrName.c_str(), (int)strlen (itrName.c_str()));
         appsToDelete.push_back(itrName);
     }
     //If empty list is passed, dynamic cache is cleared. This will clear static list also
